rag_service/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
PyYAML==6.0.2
qdrant-client[fastembed]
python-multipart==0.0.9
